[REFACTOR] Merge corkage related spring bean config to SpringConfig file SpringConfig 파일에 CorkageStore, CorkageInfo 에 대한 Repository, service 스프링 빈 등록
