+++ src/main/java/com/atable/alcholknowledge/SpringConfig.java
@@ -1,7 +1,8 @@
 package com.atable.alcholknowledge;
 
-import com.atable.alcholknowledge.repository.MemoryWineInfoReository;
-import com.atable.alcholknowledge.repository.WineInfoRepository;
+import com.atable.alcholknowledge.repository.*;
+import com.atable.alcholknowledge.service.CorkageInfoService;
+import com.atable.alcholknowledge.service.CorkageStoreService;
 import com.atable.alcholknowledge.service.WineInfoService;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -15,5 +16,25 @@ public class SpringConfig {
     @Bean
     public WineInfoRepository wineInfoRepository(){
         return new MemoryWineInfoReository();
+    }
+
+    /* Corkage Related Beans */
+    @Bean
+    public CorkageInfoService corkageInfoService() {
+        return new CorkageInfoService(corkageInfoRepository(), corkageStoreRepository());
+    }
+
+    @Bean
+    public CorkageStoreService corkageStoreService() {
+        return new CorkageStoreService(corkageStoreRepository());
+    }
+    @Bean
+    public CorkageInfoRepository corkageInfoRepository() {
+        return new MemoryCkInfoRepository();
+    }
+
+    @Bean
+    public CorkageStoreRepository corkageStoreRepository() {
+        return new MemoryCkStoreRepository();
     }
 }
